Experimental IntGridValue changes to account for DualGrid system

+++ KrasCore.Mosaic.Data/Data/Commands/SetCommand.cs
@@ -5,6 +5,6 @@ namespace KrasCore.Mosaic.Data
     public struct SetCommand
     {
         public int2 Position;
-        public int IntGridValue;
+        public short IntGridValue;
     }
 }
